Refactor and enhance contracts with improved configurations and onboarding support
- Updated `Cargo.toml` to allow non-camel case types and simplified Clippy lints.
- Enhanced `FeeOptimizationConfig` struct with `Debug` trait for better logging.
- Improved `get_provider_earnings_leaderboard` function to use indexed access for provider earnings.
- Adjusted revenue share ledger sequence type for consistency.
- Updated rate limits in `BridgeRateLimits` for better clarity and accuracy.
- Introduced onboarding module and updated related storage keys in `user_portfolio`.
- Refined event publishing in onboarding status updates for better structure.

+++ stellar-swipe/contracts/fee_collector/src/lib.rs
@@ -858,7 +858,7 @@ impl FeeCollector {
 
         let pool_amount = storage::get_revenue_share_pool(&env, &token);
         if pool_amount > 0 {
-            let ledger = env.ledger().sequence();
+            let ledger = env.ledger().sequence() as u64;
             storage::set_last_revenue_share_snapshot(&env, ledger);
 
             // Emit RevenueShareDistributed event
